# Loading module & helpers in code stack instead of manually importing in HTML #185
# What Does This Fix?
This allows the module to be loaded in both the nodejs scope and in the browser scope to. It does this by piping all objects & API's through a single`index.js` file . From here it can be used in NodeJs (intellisense etc) but to work in the browser it is bundled into a single JavaScript file through the `webpack` compiler.

# Changes
- Added basic build scripts to faciltiate the deisred outcome
- Piped everything through a single `index.js` script so that it can be loaded into nodeJS and be compiled by `webpack` to be used in the web browser scope.
- Changed the name of the source file loaded into the browser scope. To change this alter the `webpack.conig.js`  `library` & `filename` properties.
- Updated `README.md` to reflect these changes.
- Moved `NNC` files into a `worker` directory as they are inteneed to be loaded in as `web-workers` anyway

# Notes
- Will require all examples to be updated to match this namespace change